requirements.txt
anyio==3.7.1
certifi==2025.1.31
click==8.1.8
distro==1.9.0
fastapi==0.95.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
openai==0.27.8
pydantic==1.10.21
python-dotenv==0.20.0
sniffio==1.3.1
starlette==0.26.1
tqdm==4.67.1
typing_extensions==4.12.2
uvicorn==0.22.0
